HHH-15634 Fix naming of LazyBasicFieldAccessTest/LazyBasicPropertyAccessTest
For some reason they were reversed.
